requirements.txt
asgiref==3.7.2
Django==4.2.10
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
pillow==10.2.0
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
